+++ py-polars/polars/lazy/expr.py
@@ -5,7 +5,7 @@ from typing import Any, Callable, Optional, Sequence, Type, Union
 import polars as pl
 
 from ..datatypes import Boolean, DataType, Date32, Date64, Float64, Int64, Utf8
-from .functions import UDF, col, lit
+from .expr_functions import UDF, col, lit
 
 try:
     from ..polars import PyExpr
